libhwc: set window position function is moved from prepare to set function.
If window position is set in the prepare function, there is a possibility of
black screen or distrotion during video playback or camera preview (because the frame for
the new window configuration is prepared in set function. prepare function has the old frame data).

So, it's better to set the window position after having the proper frame data for the new window configuration.

Change-Id: I9507f6353c83c15dcd56bc9fe398c71407b1457c
Signed-off-by: meka rama <v.meka@samsung.com>