libhdmi: lingering UI layer bug fix
Fixed a race condition where when playing a video, the UI layer would
sometimes remain on screen.

Use two locks, one per layer, to prevent disabling a layer if there's
stuff left to output for that layer.

This solution works for now, but we should probably consider whether
it's optimal.

Change-Id: Ia74ffdbc0f69b56d9f36b9309ba64a29fa24d243
Signed-off-by: Jason Lee <aznair.lee@samsung.com>