libgscaler: When gsc_mode is GSC_CAPTURE_MODE, Change rgb-type to NARROW.
Framebuffer image in FIMD is set to NARROW and
should not be changed. But when data is transferred from
FIMD to GSC, OUT_RGB type of GSC is set to WIDE. Therefore
the black areas appear washed out because the expressed range
of colors is wider than the original range of colors.

When using the GSC, value of V4L2_CID_CSC_RANGE is set
to GSC_RGB_CSC_NARROW using v4l2_s_ctrl.

Change-Id: I3b69fbe4bbd1dbd68ca1dba204c8a75011f619d2
Signed-off-by: Lee Hwangjae <hj-yo.lee@samsung.com>