libhdmi: Change rgp-type to NARROW, when using the gscaler.
Framebuffer image in FIMD is set to NARROW and
should not be changed. But when data is transferred from
FIMD to GSC, OUT_RGB type of GSC is set to WIDE. Therefore
the black areas appear washed out because the expressed range
of colors is wider than the original range of colors.

When using the GSC, value of V4L2_CID_CSC_RANGE is set
to GSC_RGB_CSC_NARROW using v4l2_s_ctrl.

Change-Id: I881519d4d0d95cb961be7e9a1e10368971537ad6
Signed-off-by: Lee Hwangjae <hj-yo.lee@samsung.com>